Updating community.general collection for brew cask fixes
Brew Cask has switched to using `brew install --cask` instead of the old
format. There is no released collection that supports this, so we have
to pull in a hard coded commit instead.

Note: There's no trivial way to upgrade collections without a reinstall,
so it's necessary to force an install with the -f flag.